split out already written components to new files
