+++ src/__swaps__/screens/Swap/Swap.tsx
@@ -1,35 +1,18 @@
 /* eslint-disable no-nested-ternary */
-import { BlurView } from '@react-native-community/blur';
 import MaskedView from '@react-native-masked-view/masked-view';
 import c from 'chroma-js';
-import React, { ReactNode, useMemo } from 'react';
-import { StyleSheet, Text as RNText, ScrollView, TextInput, ViewStyle, StatusBar, Pressable } from 'react-native';
+import React from 'react';
+import { StyleSheet, Text as RNText, StatusBar, Pressable } from 'react-native';
 import { PanGestureHandler } from 'react-native-gesture-handler';
 import { getSoftMenuBarHeight } from 'react-native-extra-dimensions-android';
-import LinearGradient from 'react-native-linear-gradient';
-import Animated, {
-  AnimateStyle,
-  SharedValue,
-  interpolate,
-  interpolateColor,
-  runOnUI,
-  useAnimatedStyle,
-  useDerivedValue,
-  useSharedValue,
-  withSpring,
-  withTiming,
-} from 'react-native-reanimated';
+import Animated, { runOnUI, useAnimatedStyle, useDerivedValue, useSharedValue } from 'react-native-reanimated';
 import { ScreenCornerRadius } from 'react-native-screen-corner-radius';
-import { Address } from 'viem';
 
-import SwapSpinner from '@/__swaps__/assets/swapSpinner.png';
 import { ButtonPressAnimation } from '@/components/animations';
-import { AnimatedSpinner, spinnerExitConfig } from '@/__swaps__/components/animations/AnimatedSpinner';
 import { ContactAvatar } from '@/components/contacts';
 import ImageAvatar from '@/components/contacts/ImageAvatar';
 import { Page } from '@/components/layout';
-import { Input } from '@/components/inputs';
-import { Navbar, navbarHeight } from '@/components/navbar/Navbar';
+import { Navbar } from '@/components/navbar/Navbar';
 import { SheetGestureBlocker } from '@/components/sheet/SheetGestureBlocker';
 import {
   AnimatedText,
@@ -37,27 +20,19 @@ import {
   Box,
   Column,
   Columns,
-  HitSlop,
   IconContainer,
-  Inline,
   Inset,
   Separator,
   Stack,
   Text,
-  TextIcon,
   globalColors,
   useColorMode,
   useForegroundColor,
 } from '@/design-system';
-import { TextColor } from '@/design-system/color/palettes';
-import { TextWeight } from '@/design-system/components/Text/Text';
-import { TextSize } from '@/design-system/typography/typeHierarchy';
-import { useAccountAccentColor, useAccountProfile, useDimensions } from '@/hooks';
+import { useAccountProfile } from '@/hooks';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation';
 import Routes from '@/navigation/routesNames';
-import { Network } from '@/networks/types';
-import { colors } from '@/styles';
 import { useTheme } from '@/theme';
 import { safeAreaInsetValues } from '@/utils';
 
@@ -68,14 +43,17 @@ import { SwapNumberPad } from './components/SwapNumberPad';
 import { SwapSlider } from './components/SwapSlider';
 import { GasButton } from './components/GasButton';
 import { FadeMask } from './components/FadeMask';
+import { SwapBackground } from './components/SwapBackground';
+import { SwapInput } from './components/SwapInput';
+import { BalanceBadge } from './components/BalanceBadge';
+import { FlipButton } from './components/FlipButton';
+import { TokenList } from './components/TokenList';
+import { ExchangeRateBubble } from './components/ExchangeRateBubble';
 import {
-  BASE_INPUT_HEIGHT,
   BASE_INPUT_WIDTH,
   ETH_COLOR,
   ETH_COLOR_DARK,
   ETH_COLOR_DARK_ACCENT,
-  EXPANDED_INPUT_HEIGHT,
-  FOCUSED_INPUT_HEIGHT,
   INITIAL_SLIDER_POSITION,
   INPUT_INNER_WIDTH,
   INPUT_PADDING,
@@ -85,12 +63,8 @@ import {
   SLIDER_HEIGHT,
   SLIDER_WIDTH,
   THICK_BORDER_WIDTH,
-  fadeConfig,
-  springConfig,
 } from './constants';
 import {
-  DAI_ADDRESS,
-  ETH_ADDRESS,
   INPUT_ADDRESS,
   INPUT_ASSET_BALANCE,
   INPUT_ASSET_USD_PRICE,
@@ -101,11 +75,14 @@ import {
   OUTPUT_COLOR,
   OUTPUT_NETWORK,
   OUTPUT_SYMBOL,
-  USDC_ADDRESS,
 } from './dummyValues';
-import { useAnimatedSwapStyles, useSwapInputsController, useSwapNavigation, useSwapTextStyles } from './hooks/swapHooks';
+import { useAnimatedSwapStyles } from './hooks/useAnimatedSwapStyles';
+import { useSwapInputsController } from './hooks/useSwapInputsController';
+import { useSwapNavigation } from './hooks/useSwapNavigation';
+import { useSwapTextStyles } from './hooks/useSwapTextStyles';
+
 import { inputKeys } from './types';
-import { getHighContrastColor, getTintedBackgroundColor, opacity } from './utils';
+import { getHighContrastColor, opacity } from './utils';
 import { IS_ANDROID, IS_IOS } from '@/env';
 
 /** README
@@ -677,807 +654,7 @@ export function SwapScreen() {
   );
 }
 
-// TODO: Should move the components below to their own files, but some of the
-// wrappers and code in SwapScreen above should be moved out to components, so
-// leaving it all here for now
-
-const SwapBackground = ({
-  bottomColor,
-  children,
-  topColor,
-}: {
-  bottomColor: string | undefined;
-  children: ReactNode;
-  topColor: string | undefined;
-}) => {
-  const { height: deviceHeight, width: deviceWidth } = useDimensions();
-  const { isDarkMode } = useColorMode();
-
-  const fallbackColor = isDarkMode ? ETH_COLOR_DARK : ETH_COLOR;
-
-  const bottomColorDarkened = useMemo(
-    () => getTintedBackgroundColor(bottomColor || fallbackColor, isDarkMode),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [bottomColor, isDarkMode]
-  );
-  const topColorDarkened = useMemo(
-    () => getTintedBackgroundColor(topColor || fallbackColor, isDarkMode),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [isDarkMode, topColor]
-  );
-
-  return (
-    <Box
-      alignItems="center"
-      as={LinearGradient}
-      borderRadius={IS_ANDROID ? 20 : ScreenCornerRadius}
-      colors={[topColorDarkened, bottomColorDarkened]}
-      end={{ x: 0.5, y: 1 }}
-      height={{ custom: deviceHeight + (IS_ANDROID ? 24 : 0) }}
-      justifyContent="center"
-      paddingTop={{ custom: safeAreaInsetValues.top + (navbarHeight - 12) }}
-      start={{ x: 0.5, y: 0 }}
-      style={{ backgroundColor: topColor }}
-      width={{ custom: deviceWidth }}
-    >
-      {children}
-    </Box>
-  );
-};
-
-const useSwapInputStyles = ({
-  bottomInput,
-  color,
-  otherInputProgress,
-  progress,
-}: {
-  bottomInput: boolean | undefined;
-  color: string;
-  otherInputProgress: Animated.SharedValue<number>;
-  progress: Animated.SharedValue<number>;
-}) => {
-  const { isDarkMode } = useColorMode();
-
-  const { bgColor, expandedBgColor, strokeColor, expandedStrokeColor, mixedShadowColor } = useMemo(() => {
-    const bgColor = isDarkMode ? opacity(color, 0.08) : opacity(globalColors.white100, 0.8);
-    const expandedBgColor = isDarkMode ? bgColor : opacity(globalColors.white100, 0.8);
-    const strokeColor = isDarkMode ? opacity(color === ETH_COLOR_DARK ? ETH_COLOR_DARK_ACCENT : color, 0.06) : globalColors.white100;
-    const expandedStrokeColor = isDarkMode ? opacity(color, 0.1) : globalColors.white100;
-    const mixedShadowColor = isDarkMode ? 'transparent' : c.mix(color, globalColors.grey100, 0.84).hex();
-
-    return {
-      bgColor,
-      expandedBgColor,
-      strokeColor,
-      expandedStrokeColor,
-      mixedShadowColor,
-    };
-  }, [color, isDarkMode]);
-
-  const containerStyle = useAnimatedStyle(() => {
-    return {
-      opacity: otherInputProgress.value === 2 ? withTiming(0, fadeConfig) : withTiming(1, fadeConfig),
-      transform: [
-        {
-          translateY:
-            progress.value === 2
-              ? bottomInput
-                ? withSpring(-191, springConfig)
-                : withSpring(-77, springConfig)
-              : withSpring(0, springConfig),
-        },
-      ],
-    };
-  }, [bottomInput, otherInputProgress, progress]);
-
-  const inputStyle = useAnimatedStyle(() => {
-    return {
-      backgroundColor: withTiming(interpolateColor(progress.value, [0, 1], [bgColor, expandedBgColor]), fadeConfig),
-      borderColor: withTiming(interpolateColor(progress.value, [0, 1], [strokeColor, expandedStrokeColor]), fadeConfig),
-      height: withSpring(
-        interpolate(progress.value, [0, 1, 2], [BASE_INPUT_HEIGHT, EXPANDED_INPUT_HEIGHT, FOCUSED_INPUT_HEIGHT], 'clamp'),
-        springConfig
-      ),
-      transform: [
-        {
-          translateY: bottomInput
-            ? withSpring(
-                interpolate(otherInputProgress.value, [0, 1, 2], [0, 0, EXPANDED_INPUT_HEIGHT - FOCUSED_INPUT_HEIGHT], 'clamp'),
-                springConfig
-              )
-            : 0,
-        },
-      ],
-    };
-  }, [bottomInput, otherInputProgress, progress, bgColor, expandedBgColor, strokeColor, expandedStrokeColor]);
-
-  return { containerStyle, inputStyle, mixedShadowColor };
-};
-
-export const SwapInput = ({
-  children,
-  color,
-  bottomInput,
-  otherInputProgress,
-  progress,
-}: {
-  children?: ReactNode;
-  color: string | undefined;
-  bottomInput?: boolean;
-  otherInputProgress: SharedValue<number>;
-  progress: SharedValue<number>;
-}) => {
-  const { isDarkMode } = useColorMode();
-
-  const colorWithFallback = useMemo(() => color || (isDarkMode ? ETH_COLOR_DARK : ETH_COLOR), [color, isDarkMode]);
-
-  const { inputStyle, containerStyle, mixedShadowColor } = useSwapInputStyles({
-    bottomInput,
-    color: colorWithFallback,
-    otherInputProgress,
-    progress,
-  });
-
-  return (
-    <Box
-      as={Animated.View}
-      style={[containerStyle, styles.staticInputContainerStyles, { shadowColor: mixedShadowColor }]}
-      width={{ custom: BASE_INPUT_WIDTH }}
-    >
-      <Box as={Animated.View} style={[inputStyle, styles.staticInputStyles]}>
-        {children}
-      </Box>
-    </Box>
-  );
-};
-
-const BalanceBadge = ({ color, label, weight }: { color?: TextColor; label: string; weight?: TextWeight }) => {
-  const { isDarkMode } = useColorMode();
-
-  return (
-    <Bleed vertical={{ custom: 5.5 }}>
-      <Box
-        alignItems="center"
-        borderRadius={8.5}
-        height={{ custom: 20 }}
-        justifyContent="center"
-        paddingHorizontal={{ custom: 5 }}
-        style={{
-          borderColor: isDarkMode ? SEPARATOR_COLOR : LIGHT_SEPARATOR_COLOR,
-          borderWidth: THICK_BORDER_WIDTH,
-        }}
-      >
-        <Text
-          align="center"
-          color={color || 'labelQuaternary'}
-          size="13pt"
-          style={{
-            opacity: label === 'No Balance' ? (isDarkMode ? 0.6 : 0.75) : undefined,
-          }}
-          weight={weight || 'bold'}
-        >
-          {label}
-        </Text>
-      </Box>
-    </Bleed>
-  );
-};
-
-const AnimatedBlurView = Animated.createAnimatedComponent(BlurView);
-
-const FlipButton = ({
-  bottomColor,
-  flipButtonStyle,
-  focusedSearchStyle,
-  isFetching,
-}: {
-  bottomColor: string;
-  flipButtonStyle: AnimateStyle<ViewStyle>;
-  focusedSearchStyle: AnimateStyle<ViewStyle>;
-  isFetching: boolean;
-}) => {
-  const { isDarkMode } = useColorMode();
-
-  const fetchingStyle = useAnimatedStyle(() => {
-    return {
-      borderWidth: isFetching ? withTiming(2, { duration: 300 }) : withTiming(THICK_BORDER_WIDTH, spinnerExitConfig),
-    };
-  });
-
-  return (
-    <Box
-      alignItems="center"
-      as={Animated.View}
-      justifyContent="center"
-      style={[flipButtonStyle, focusedSearchStyle, { height: 12, width: 28, zIndex: 10 }]}
-    >
-      <Box
-        style={{
-          shadowColor: isDarkMode ? globalColors.grey100 : c.mix(bottomColor, colors.dark, 0.84).hex(),
-          shadowOffset: {
-            width: 0,
-            height: isDarkMode ? 4 : 4,
-          },
-          elevation: 8,
-          shadowOpacity: isDarkMode ? 0.3 : 0.1,
-          shadowRadius: isDarkMode ? 6 : 8,
-        }}
-      >
-        <ButtonPressAnimation scaleTo={0.8} style={{ paddingHorizontal: 20, paddingVertical: 8 }}>
-          {/* TODO: Temp fix - rewrite to actually avoid type errors */}
-          {/* @ts-expect-error The conditional as={} is causing type errors */}
-          <Box
-            alignItems="center"
-            as={IS_IOS ? AnimatedBlurView : Animated.View}
-            justifyContent="center"
-            style={[
-              fetchingStyle,
-              styles.flipButton,
-              {
-                backgroundColor: IS_ANDROID ? (isDarkMode ? globalColors.blueGrey100 : globalColors.white100) : undefined,
-                borderColor: isDarkMode ? SEPARATOR_COLOR : opacity(globalColors.white100, 0.5),
-              },
-            ]}
-            // eslint-disable-next-line react/jsx-props-no-spreading
-            {...(IS_IOS && {
-              blurAmount: 10,
-              blurType: isDarkMode ? undefined : 'light',
-            })}
-          >
-            <IconContainer size={24} opacity={isDarkMode ? 0.6 : 0.8}>
-              <Box alignItems="center" justifyContent="center">
-                <Bleed bottom={{ custom: IS_IOS ? 0.5 : 4 }}>
-                  <Text align="center" color="labelTertiary" size="icon 13px" weight="heavy">
-                    􀆈
-                  </Text>
-                </Bleed>
-              </Box>
-            </IconContainer>
-          </Box>
-        </ButtonPressAnimation>
-      </Box>
-      <Box pointerEvents="none" position="absolute">
-        <AnimatedSpinner color={bottomColor} isLoading={isFetching} scaleInFrom={1} size={32} src={SwapSpinner} />
-      </Box>
-    </Box>
-  );
-};
-
-const TokenList = ({
-  color,
-  handleExitSearch,
-  handleFocusSearch,
-  isFocused,
-  output,
-  setIsFocused,
-}: {
-  color: string;
-  handleExitSearch: () => void;
-  handleFocusSearch: () => void;
-  isFocused: boolean;
-  output?: boolean;
-  setIsFocused: React.Dispatch<React.SetStateAction<boolean>>;
-}) => {
-  const { accentColor: accountColor } = useAccountAccentColor();
-  const { isDarkMode } = useColorMode();
-  const { width: deviceWidth } = useDimensions();
-  const theme = useTheme();
-
-  const blue = useForegroundColor('blue');
-  const red = useForegroundColor('red');
-
-  const accentColor = useMemo(() => {
-    if (c.contrast(accountColor, isDarkMode ? '#191A1C' : globalColors.white100) < (isDarkMode ? 2.125 : 1.5)) {
-      const shiftedColor = isDarkMode ? c(accountColor).brighten(1).saturate(0.5).css() : c(accountColor).darken(0.5).saturate(0.5).css();
-      return shiftedColor;
-    } else {
-      return accountColor;
-    }
-  }, [accountColor, isDarkMode]);
-
-  return (
-    <Stack>
-      <Stack space="20px">
-        <SearchInput
-          color={color}
-          handleExitSearch={handleExitSearch}
-          handleFocusSearch={handleFocusSearch}
-          isFocused={isFocused}
-          output={output}
-          setIsFocused={setIsFocused}
-        />
-        <Separator color="separatorTertiary" thickness={1} />
-      </Stack>
-      <ScrollView
-        contentContainerStyle={{
-          paddingBottom: isFocused ? EXPANDED_INPUT_HEIGHT - FOCUSED_INPUT_HEIGHT + 20 : 20,
-          paddingTop: 20,
-        }}
-        showsVerticalScrollIndicator={false}
-        style={{
-          alignSelf: 'center',
-          height: EXPANDED_INPUT_HEIGHT - 77,
-          paddingHorizontal: 20,
-          width: deviceWidth - 24,
-        }}
-      >
-        <Stack space={output ? '28px' : '20px'}>
-          <Stack space="20px">
-            <Inline alignHorizontal="justify" alignVertical="center">
-              {output ? (
-                <Inline alignVertical="center" space="6px">
-                  <Bleed vertical="4px">
-                    <Box alignItems="center" justifyContent="center" marginBottom={{ custom: -0.5 }} width={{ custom: 16 }}>
-                      <Bleed space={isDarkMode ? '16px' : undefined}>
-                        <RNText
-                          style={
-                            isDarkMode
-                              ? [
-                                  styles.textIconGlow,
-                                  {
-                                    textShadowColor: opacity(red, 0.28),
-                                  },
-                                ]
-                              : undefined
-                          }
-                        >
-                          <Text align="center" color="red" size="icon 13px" weight="heavy">
-                            􀙬
-                          </Text>
-                        </RNText>
-                      </Bleed>
-                    </Box>
-                  </Bleed>
-                  <Text color="label" size="15pt" weight="heavy">
-                    Trending
-                  </Text>
-                </Inline>
-              ) : (
-                <Inline alignVertical="center" space="6px">
-                  <Bleed vertical="4px">
-                    <Box alignItems="center" justifyContent="center" width={{ custom: 18 }}>
-                      <Bleed space={isDarkMode ? '16px' : undefined}>
-                        <RNText
-                          style={
-                            isDarkMode
-                              ? [
-                                  styles.textIconGlow,
-                                  {
-                                    textShadowColor: opacity(accentColor, 0.2),
-                                  },
-                                ]
-                              : undefined
-                          }
-                        >
-                          <Text align="center" color={{ custom: accentColor }} size="icon 13px" weight="black">
-                            􀣽
-                          </Text>
-                        </RNText>
-                      </Bleed>
-                    </Box>
-                  </Bleed>
-                  <Text color="label" size="15pt" weight="heavy">
-                    My Tokens
-                  </Text>
-                </Inline>
-              )}
-              <ButtonPressAnimation>
-                <HitSlop space="10px">
-                  <Inline alignVertical="center" space="6px" wrap={false}>
-                    {output && <SwapCoinIcon address={ETH_ADDRESS} network={Network.mainnet} small symbol="ETH" theme={theme} />}
-                    <Text align="right" color={isDarkMode ? 'labelSecondary' : 'label'} size="15pt" weight="heavy">
-                      {output ? 'Ethereum' : 'All Networks'}
-                    </Text>
-                    <Text align="center" color={isDarkMode ? 'labelTertiary' : 'labelSecondary'} size="icon 13px" weight="bold">
-                      􀆏
-                    </Text>
-                  </Inline>
-                </HitSlop>
-              </ButtonPressAnimation>
-            </Inline>
-            <CoinRow
-              address={ETH_ADDRESS}
-              balance="7"
-              isTrending={output}
-              name="Ethereum"
-              nativeBalance="$18,320"
-              output={output}
-              symbol="ETH"
-            />
-            <CoinRow
-              address={USDC_ADDRESS}
-              balance="2,640"
-              isTrending={output}
-              name="USD Coin"
-              nativeBalance="$2,640"
-              output={output}
-              symbol="USDC"
-            />
-            <CoinRow
-              address={DAI_ADDRESS}
-              balance="2,800.02"
-              isTrending={output}
-              name="Dai"
-              nativeBalance="$2,800"
-              output={output}
-              symbol="DAI"
-            />
-          </Stack>
-          <Stack space="20px">
-            {output && (
-              <Inline alignVertical="center" space="6px">
-                <Bleed vertical="4px">
-                  <Box alignItems="center" justifyContent="center" width={{ custom: 16 }}>
-                    <Bleed space={isDarkMode ? '16px' : undefined}>
-                      <RNText
-                        style={
-                          isDarkMode
-                            ? [
-                                styles.textIconGlow,
-                                {
-                                  textShadowColor: opacity(blue, 0.28),
-                                },
-                              ]
-                            : undefined
-                        }
-                      >
-                        <Text align="center" color="blue" size="icon 13px" weight="heavy">
-                          􀐫
-                        </Text>
-                      </RNText>
-                    </Bleed>
-                  </Box>
-                </Bleed>
-                <Text color="label" size="15pt" weight="heavy">
-                  Recent
-                </Text>
-              </Inline>
-            )}
-            <CoinRow
-              address="0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9"
-              balance="428.25"
-              name="Aave"
-              nativeBalance="$1,400"
-              output={output}
-              symbol="AAVE"
-            />
-            <CoinRow
-              address="0x2260fac5e5542a773aa44fbcfedf7c193bc2c599"
-              balance="0.042819"
-              name="Wrapped Bitcoin"
-              nativeBalance="$1,025"
-              output={output}
-              symbol="WBTC"
-            />
-            <CoinRow
-              address="0xc00e94cb662c3520282e6f5717214004a7f26888"
-              balance="72.2806"
-              name="Compound"
-              nativeBalance="$350.04"
-              output={output}
-              symbol="COMP"
-            />
-            <CoinRow
-              address="0x1f9840a85d5af5bf1d1762f925bdaddc4201f984"
-              balance="62.82"
-              name="Uniswap"
-              nativeBalance="$289.52"
-              output={output}
-              symbol="UNI"
-            />
-            <CoinRow
-              address="0x514910771af9ca656af840dff83e8264ecf986ca"
-              balance="27.259"
-              name="Chainlink"
-              nativeBalance="$87.50"
-              output={output}
-              symbol="LINK"
-            />
-          </Stack>
-        </Stack>
-      </ScrollView>
-    </Stack>
-  );
-};
-
-const SearchInput = ({
-  color,
-  handleExitSearch,
-  handleFocusSearch,
-  isFocused,
-  output,
-  setIsFocused,
-}: {
-  color: string;
-  handleExitSearch: () => void;
-  handleFocusSearch: () => void;
-  isFocused?: boolean;
-  output?: boolean;
-  setIsFocused: React.Dispatch<React.SetStateAction<boolean>>;
-}) => {
-  const { isDarkMode } = useColorMode();
-
-  const inputRef = React.useRef<TextInput>(null);
-  const [query, setQuery] = React.useState('');
-
-  const fillTertiary = useForegroundColor('fillTertiary');
-  const label = useForegroundColor('label');
-  const labelQuaternary = useForegroundColor('labelQuaternary');
-
-  return (
-    <Box width="full">
-      <Columns alignHorizontal="justify" alignVertical="center" space="20px">
-        <Box>
-          <Bleed horizontal="8px" vertical="24px">
-            <Box paddingHorizontal="8px" paddingVertical="20px">
-              <Columns alignVertical="center" space="10px">
-                <Column width="content">
-                  <Box
-                    alignItems="center"
-                    borderRadius={18}
-                    height={{ custom: 36 }}
-                    justifyContent="center"
-                    style={{
-                      backgroundColor: isDarkMode ? 'transparent' : opacity(fillTertiary, 0.03),
-                      borderColor: isDarkMode ? SEPARATOR_COLOR : opacity(LIGHT_SEPARATOR_COLOR, 0.01),
-                      borderWidth: THICK_BORDER_WIDTH,
-                    }}
-                    width={{ custom: 36 }}
-                  >
-                    <Text align="center" color="labelQuaternary" size="icon 17px" weight="bold">
-                      􀊫
-                    </Text>
-                  </Box>
-                </Column>
-                <Input
-                  onBlur={() => {
-                    handleExitSearch();
-                    setIsFocused(false);
-                  }}
-                  onChange={(value: string) => setQuery(value)}
-                  onFocus={() => {
-                    handleFocusSearch();
-                    setIsFocused(true);
-                  }}
-                  placeholder={output ? 'Find a token to buy' : 'Search your tokens'}
-                  placeholderTextColor={isDarkMode ? opacity(labelQuaternary, 0.3) : labelQuaternary}
-                  ref={inputRef}
-                  selectionColor={color}
-                  spellCheck={false}
-                  style={{
-                    color: label,
-                    fontSize: 17,
-                    fontWeight: 'bold',
-                    height: 44,
-                    zIndex: 10,
-                  }}
-                  value={query}
-                />
-              </Columns>
-            </Box>
-          </Bleed>
-        </Box>
-        <Column width="content">
-          <ButtonPressAnimation
-            onPress={() => {
-              if (!isFocused) {
-                handleExitSearch();
-              } else {
-                inputRef.current?.blur();
-                setIsFocused(false);
-              }
-            }}
-            scaleTo={0.8}
-          >
-            <Box
-              alignItems="center"
-              borderRadius={18}
-              height={{ custom: 36 }}
-              justifyContent="center"
-              paddingHorizontal={{ custom: 12 - THICK_BORDER_WIDTH }}
-              style={{
-                backgroundColor: opacity(color, isDarkMode ? 0.1 : 0.08),
-                borderColor: opacity(color, isDarkMode ? 0.06 : 0.01),
-                borderWidth: THICK_BORDER_WIDTH,
-              }}
-            >
-              <Text align="center" color={{ custom: color }} size="17pt" weight="heavy">
-                {isFocused ? 'Cancel' : 'Close'}
-              </Text>
-            </Box>
-          </ButtonPressAnimation>
-        </Column>
-      </Columns>
-    </Box>
-  );
-};
-
-const CoinRow = ({
-  address,
-  balance,
-  isTrending,
-  name,
-  nativeBalance,
-  onPress,
-  output,
-  symbol,
-}: {
-  address: Address | 'eth';
-  balance: string;
-  isTrending?: boolean;
-  name: string;
-  nativeBalance: string;
-  onPress?: () => void;
-  output?: boolean;
-  symbol: string;
-}) => {
-  const theme = useTheme();
-
-  const percentChange = useMemo(() => {
-    if (isTrending) {
-      const rawChange = Math.random() * 30;
-      const isNegative = Math.random() < 0.2;
-      const prefix = isNegative ? '-' : '+';
-      const color: TextColor = isNegative ? 'red' : 'green';
-      const change = `${rawChange.toFixed(1)}%`;
-
-      return { change, color, prefix };
-    }
-  }, [isTrending]);
-
-  return (
-    <ButtonPressAnimation onPress={onPress} scaleTo={0.95}>
-      <HitSlop vertical="10px">
-        <Box alignItems="center" flexDirection="row" justifyContent="space-between" width="full">
-          <Inline alignVertical="center" space="10px">
-            <SwapCoinIcon address={address} large network={Network.mainnet} symbol={symbol} theme={theme} />
-            <Stack space="10px">
-              <Text color="label" size="17pt" weight="semibold">
-                {name}
-              </Text>
-              <Inline alignVertical="center" space={{ custom: 5 }}>
-                <Text color="labelTertiary" size="13pt" weight="semibold">
-                  {output ? symbol : `${balance} ${symbol}`}
-                </Text>
-                {isTrending && percentChange && (
-                  <Inline alignVertical="center" space={{ custom: 1 }}>
-                    <Text align="center" color={percentChange.color} size="12pt" weight="bold">
-                      {percentChange.prefix}
-                    </Text>
-                    <Text color={percentChange.color} size="13pt" weight="semibold">
-                      {percentChange.change}
-                    </Text>
-                  </Inline>
-                )}
-              </Inline>
-            </Stack>
-          </Inline>
-          {output ? (
-            <Inline space="8px">
-              <CoinRowButton icon="􀅳" outline size="icon 14px" />
-              <CoinRowButton icon="􀋃" weight="black" />
-            </Inline>
-          ) : (
-            <BalancePill balance={nativeBalance} />
-          )}
-        </Box>
-      </HitSlop>
-    </ButtonPressAnimation>
-  );
-};
-
-const BalancePill = ({ balance }: { balance: string }) => {
-  const { isDarkMode } = useColorMode();
-
-  return (
-    <Box
-      alignItems="center"
-      borderRadius={14}
-      height={{ custom: 28 }}
-      justifyContent="center"
-      paddingHorizontal={{ custom: 8 - THICK_BORDER_WIDTH }}
-      style={{
-        backgroundColor: 'transparent',
-        borderColor: isDarkMode ? SEPARATOR_COLOR : LIGHT_SEPARATOR_COLOR,
-        borderCurve: 'continuous',
-        borderWidth: THICK_BORDER_WIDTH,
-        overflow: 'hidden',
-      }}
-    >
-      <Text align="center" color="labelTertiary" size="15pt" weight="bold">
-        {balance}
-      </Text>
-    </Box>
-  );
-};
-
-const ExchangeRateBubble = () => {
-  const { isDarkMode } = useColorMode();
-
-  const fillTertiary = useForegroundColor('fillTertiary');
-
-  return (
-    <Box
-      alignItems="center"
-      borderRadius={15}
-      height={{ custom: 30 }}
-      justifyContent="center"
-      paddingHorizontal="10px"
-      style={{ borderColor: isDarkMode ? SEPARATOR_COLOR : LIGHT_SEPARATOR_COLOR, borderWidth: THICK_BORDER_WIDTH }}
-    >
-      <Inline alignHorizontal="center" alignVertical="center" space="6px" wrap={false}>
-        <Text align="center" color="labelQuaternary" size="13pt" style={{ opacity: isDarkMode ? 0.6 : 0.75 }} weight="heavy">
-          1 ETH
-        </Text>
-        <Box
-          borderRadius={10}
-          height={{ custom: 20 }}
-          paddingTop={{ custom: 0.25 }}
-          style={{ backgroundColor: opacity(fillTertiary, 0.04) }}
-          width={{ custom: 20 }}
-        >
-          <TextIcon color="labelQuaternary" containerSize={20} opacity={isDarkMode ? 0.6 : 0.75} size="icon 10px" weight="heavy">
-            􀄭
-          </TextIcon>
-        </Box>
-        <Text align="center" color="labelQuaternary" size="13pt" style={{ opacity: isDarkMode ? 0.6 : 0.75 }} weight="heavy">
-          1,624.04 USDC
-        </Text>
-      </Inline>
-    </Box>
-  );
-};
-
-const CoinRowButton = ({
-  icon,
-  onPress,
-  outline,
-  size,
-  weight,
-}: {
-  icon: string;
-  onPress?: () => void;
-  outline?: boolean;
-  size?: TextSize;
-  weight?: TextWeight;
-}) => {
-  const { isDarkMode } = useColorMode();
-  const fillTertiary = useForegroundColor('fillTertiary');
-  const fillQuaternary = useForegroundColor('fillQuaternary');
-  const separatorTertiary = useForegroundColor('separatorTertiary');
-
-  return (
-    <ButtonPressAnimation onPress={onPress} scaleTo={0.8}>
-      <Box
-        alignItems="center"
-        borderRadius={14}
-        height={{ custom: 28 }}
-        justifyContent="center"
-        style={{
-          backgroundColor: outline ? 'transparent' : isDarkMode ? fillQuaternary : opacity(fillTertiary, 0.04),
-          borderColor: outline ? (isDarkMode ? SEPARATOR_COLOR : LIGHT_SEPARATOR_COLOR) : separatorTertiary,
-          borderWidth: THICK_BORDER_WIDTH,
-        }}
-        width={{ custom: 28 }}
-      >
-        <TextIcon
-          color="labelQuaternary"
-          containerSize={28}
-          opacity={isDarkMode ? 0.6 : 0.75}
-          size={size || 'icon 12px'}
-          weight={weight || 'heavy'}
-        >
-          {icon}
-        </TextIcon>
-      </Box>
-    </ButtonPressAnimation>
-  );
-};
-
-// /---- 📐 Styles 📐 ----/ //
-//
-// TODO: Should move the remaining unnecessarily inlined styles here
-const styles = StyleSheet.create({
+export const styles = StyleSheet.create({
   backgroundOverlay: {
     backgroundColor: 'rgba(0, 0, 0, 0.88)',
   },
@@ -1536,5 +713,3 @@ const styles = StyleSheet.create({
     textShadowRadius: 10,
   },
 });
-//
-// /---- END styles ----/ //
