+++ src/__swaps__/screens/Swap/components/AnimatedBlurView.tsx
@@ -1,0 +1,4 @@
+import { BlurView } from '@react-native-community/blur';
+import Animated from 'react-native-reanimated';
+
+export const AnimatedBlurView = Animated.createAnimatedComponent(BlurView);
