+++ src/__swaps__/screens/Swap/hooks/useAnimatedSwapStyles.ts
@@ -1,0 +1,91 @@
+import { SharedValue, interpolate, useAnimatedStyle, withSpring, withTiming } from 'react-native-reanimated';
+
+import { BASE_INPUT_HEIGHT, EXPANDED_INPUT_HEIGHT, FOCUSED_INPUT_HEIGHT, fadeConfig, springConfig } from '../constants';
+
+export function useAnimatedSwapStyles({
+  inputProgress,
+  outputProgress,
+}: {
+  inputProgress: SharedValue<number>;
+  outputProgress: SharedValue<number>;
+}) {
+  const flipButtonStyle = useAnimatedStyle(() => {
+    return {
+      transform: [
+        {
+          translateY: withSpring(
+            interpolate(inputProgress.value, [0, 1, 2], [0, 0, EXPANDED_INPUT_HEIGHT - FOCUSED_INPUT_HEIGHT], 'clamp'),
+            springConfig
+          ),
+        },
+      ],
+    };
+  });
+
+  const focusedSearchStyle = useAnimatedStyle(() => {
+    return {
+      opacity: inputProgress.value === 2 || outputProgress.value === 2 ? withTiming(0, fadeConfig) : withTiming(1, fadeConfig),
+      pointerEvents: inputProgress.value === 2 || outputProgress.value === 2 ? 'none' : 'auto',
+    };
+  });
+
+  const hideWhenInputsExpanded = useAnimatedStyle(() => {
+    return {
+      opacity: inputProgress.value > 0 || outputProgress.value > 0 ? withTiming(0, fadeConfig) : withTiming(1, fadeConfig),
+      pointerEvents: inputProgress.value > 0 || outputProgress.value > 0 ? 'none' : 'auto',
+    };
+  });
+
+  const inputStyle = useAnimatedStyle(() => {
+    return {
+      opacity: withTiming(interpolate(inputProgress.value, [0, 1], [1, 0], 'clamp'), fadeConfig),
+      pointerEvents: inputProgress.value === 0 ? 'auto' : 'none',
+    };
+  });
+
+  const inputTokenListStyle = useAnimatedStyle(() => {
+    return {
+      opacity: withTiming(interpolate(inputProgress.value, [0, 1], [0, 1], 'clamp'), fadeConfig),
+      pointerEvents: inputProgress.value === 0 ? 'none' : 'auto',
+    };
+  });
+
+  const keyboardStyle = useAnimatedStyle(() => {
+    const progress = Math.min(inputProgress.value + outputProgress.value, 1);
+
+    return {
+      opacity: withTiming(1 - progress, fadeConfig),
+      transform: [
+        {
+          translateY: withSpring(progress * (EXPANDED_INPUT_HEIGHT - BASE_INPUT_HEIGHT), springConfig),
+        },
+        { scale: withSpring(0.925 + (1 - progress) * 0.075, springConfig) },
+      ],
+    };
+  });
+
+  const outputStyle = useAnimatedStyle(() => {
+    return {
+      opacity: withTiming(interpolate(outputProgress.value, [0, 1], [1, 0], 'clamp'), fadeConfig),
+      pointerEvents: outputProgress.value === 0 ? 'auto' : 'none',
+    };
+  });
+
+  const outputTokenListStyle = useAnimatedStyle(() => {
+    return {
+      opacity: withTiming(interpolate(outputProgress.value, [0, 1], [0, 1], 'clamp'), fadeConfig),
+      pointerEvents: outputProgress.value === 0 ? 'none' : 'auto',
+    };
+  });
+
+  return {
+    flipButtonStyle,
+    focusedSearchStyle,
+    hideWhenInputsExpanded,
+    inputStyle,
+    inputTokenListStyle,
+    keyboardStyle,
+    outputStyle,
+    outputTokenListStyle,
+  };
+}
