+++ src/__swaps__/screens/Swap/hooks/useSwapInputsController.ts
@@ -1,41 +1,16 @@
 import React, { useCallback, useMemo, useRef } from 'react';
 import Animated, {
-  Easing,
   SharedValue,
-  interpolate,
-  interpolateColor,
   runOnJS,
   runOnUI,
   useAnimatedReaction,
-  useAnimatedStyle,
   useDerivedValue,
   useSharedValue,
-  withRepeat,
-  withSequence,
   withSpring,
-  withTiming,
 } from 'react-native-reanimated';
 import { useDebouncedCallback } from 'use-debounce';
-import { useColorMode, useForegroundColor } from '@/design-system';
 
-import {
-  BASE_INPUT_HEIGHT,
-  ETH_COLOR_DARK,
-  ETH_COLOR_DARK_ACCENT,
-  EXPANDED_INPUT_HEIGHT,
-  FOCUSED_INPUT_HEIGHT,
-  SCRUBBER_WIDTH,
-  SLIDER_COLLAPSED_HEIGHT,
-  SLIDER_HEIGHT,
-  SLIDER_WIDTH,
-  caretConfig,
-  fadeConfig,
-  pulsingConfig,
-  sliderConfig,
-  slowFadeConfig,
-  snappySpringConfig,
-  springConfig,
-} from '../constants';
+import { SCRUBBER_WIDTH, SLIDER_WIDTH, snappySpringConfig } from '../constants';
 import { IS_INPUT_STABLECOIN, IS_OUTPUT_STABLECOIN, SWAP_FEE } from '../dummyValues';
 import { inputKeys, inputMethods } from '../types';
 import {
@@ -45,102 +20,9 @@ import {
   countDecimalPlaces,
   findNiceIncrement,
   niceIncrementFormatter,
-  opacity,
   trimTrailingZeros,
   valueBasedDecimalFormatter,
 } from '../utils';
-
-// TODO: Probably cleaner to use a shared context to supply the shared values
-
-// /---- ⚙️ Swap UI hooks ⚙️ ----/ //
-//
-export function useAnimatedSwapStyles({
-  inputProgress,
-  outputProgress,
-}: {
-  inputProgress: SharedValue<number>;
-  outputProgress: SharedValue<number>;
-}) {
-  const flipButtonStyle = useAnimatedStyle(() => {
-    return {
-      transform: [
-        {
-          translateY: withSpring(
-            interpolate(inputProgress.value, [0, 1, 2], [0, 0, EXPANDED_INPUT_HEIGHT - FOCUSED_INPUT_HEIGHT], 'clamp'),
-            springConfig
-          ),
-        },
-      ],
-    };
-  });
-
-  const focusedSearchStyle = useAnimatedStyle(() => {
-    return {
-      opacity: inputProgress.value === 2 || outputProgress.value === 2 ? withTiming(0, fadeConfig) : withTiming(1, fadeConfig),
-      pointerEvents: inputProgress.value === 2 || outputProgress.value === 2 ? 'none' : 'auto',
-    };
-  });
-
-  const hideWhenInputsExpanded = useAnimatedStyle(() => {
-    return {
-      opacity: inputProgress.value > 0 || outputProgress.value > 0 ? withTiming(0, fadeConfig) : withTiming(1, fadeConfig),
-      pointerEvents: inputProgress.value > 0 || outputProgress.value > 0 ? 'none' : 'auto',
-    };
-  });
-
-  const inputStyle = useAnimatedStyle(() => {
-    return {
-      opacity: withTiming(interpolate(inputProgress.value, [0, 1], [1, 0], 'clamp'), fadeConfig),
-      pointerEvents: inputProgress.value === 0 ? 'auto' : 'none',
-    };
-  });
-
-  const inputTokenListStyle = useAnimatedStyle(() => {
-    return {
-      opacity: withTiming(interpolate(inputProgress.value, [0, 1], [0, 1], 'clamp'), fadeConfig),
-      pointerEvents: inputProgress.value === 0 ? 'none' : 'auto',
-    };
-  });
-
-  const keyboardStyle = useAnimatedStyle(() => {
-    const progress = Math.min(inputProgress.value + outputProgress.value, 1);
-
-    return {
-      opacity: withTiming(1 - progress, fadeConfig),
-      transform: [
-        {
-          translateY: withSpring(progress * (EXPANDED_INPUT_HEIGHT - BASE_INPUT_HEIGHT), springConfig),
-        },
-        { scale: withSpring(0.925 + (1 - progress) * 0.075, springConfig) },
-      ],
-    };
-  });
-
-  const outputStyle = useAnimatedStyle(() => {
-    return {
-      opacity: withTiming(interpolate(outputProgress.value, [0, 1], [1, 0], 'clamp'), fadeConfig),
-      pointerEvents: outputProgress.value === 0 ? 'auto' : 'none',
-    };
-  });
-
-  const outputTokenListStyle = useAnimatedStyle(() => {
-    return {
-      opacity: withTiming(interpolate(outputProgress.value, [0, 1], [0, 1], 'clamp'), fadeConfig),
-      pointerEvents: outputProgress.value === 0 ? 'none' : 'auto',
-    };
-  });
-
-  return {
-    flipButtonStyle,
-    focusedSearchStyle,
-    hideWhenInputsExpanded,
-    inputStyle,
-    inputTokenListStyle,
-    keyboardStyle,
-    outputStyle,
-    outputTokenListStyle,
-  };
-}
 
 export function useSwapInputsController({
   focusedInput,
@@ -554,258 +436,3 @@ export function useSwapInputsController({
     percentageToSwap,
   };
 }
-
-export function useSwapNavigation({
-  inputProgress,
-  outputProgress,
-}: {
-  inputProgress: SharedValue<number>;
-  outputProgress: SharedValue<number>;
-}) {
-  // TODO: This can all be simplified
-  const handleExitSearch = useCallback(() => {
-    'worklet';
-    if (inputProgress.value === 1) {
-      inputProgress.value = 0;
-    }
-    if (outputProgress.value === 1) {
-      outputProgress.value = 0;
-    }
-    if (inputProgress.value === 2) {
-      inputProgress.value = 1;
-    }
-    if (outputProgress.value === 2) {
-      outputProgress.value = 1;
-    }
-  }, [inputProgress, outputProgress]);
-
-  const handleFocusInputSearch = useCallback(() => {
-    'worklet';
-    if (inputProgress.value !== 2) {
-      inputProgress.value = 2;
-    }
-  }, [inputProgress]);
-
-  const handleFocusOutputSearch = useCallback(() => {
-    'worklet';
-    if (outputProgress.value !== 2) {
-      outputProgress.value = 2;
-    }
-  }, [outputProgress]);
-
-  const handleInputPress = useCallback(() => {
-    'worklet';
-    if (inputProgress.value === 0) {
-      inputProgress.value = 1;
-      outputProgress.value = 0;
-    } else {
-      inputProgress.value = 0;
-    }
-  }, [inputProgress, outputProgress]);
-
-  const handleOutputPress = useCallback(() => {
-    'worklet';
-    if (outputProgress.value === 0) {
-      outputProgress.value = 1;
-      inputProgress.value = 0;
-    } else {
-      outputProgress.value = 0;
-    }
-  }, [inputProgress, outputProgress]);
-
-  return {
-    handleExitSearch,
-    handleFocusInputSearch,
-    handleFocusOutputSearch,
-    handleInputPress,
-    handleOutputPress,
-  };
-}
-
-export function useSwapTextStyles({
-  bottomColor,
-  focusedInput,
-  inputMethod,
-  inputProgress,
-  inputValues,
-  isQuoteStale,
-  outputProgress,
-  sliderPressProgress,
-  topColor,
-}: {
-  bottomColor: string;
-  focusedInput: SharedValue<inputKeys>;
-  inputMethod: SharedValue<inputMethods>;
-  inputProgress: SharedValue<number>;
-  inputValues: SharedValue<{ [key in inputKeys]: number | string }>;
-  isQuoteStale: SharedValue<number>;
-  outputProgress: SharedValue<number>;
-  sliderPressProgress: SharedValue<number>;
-  topColor: string;
-}) {
-  const { isDarkMode } = useColorMode();
-
-  const labelSecondary = useForegroundColor('labelSecondary');
-  const labelTertiary = useForegroundColor('labelTertiary');
-  const zeroAmountColor = opacity(labelSecondary, 0.2);
-
-  const isInputStale = useDerivedValue(() => {
-    const isAdjustingOutputValue = inputMethod.value === 'outputAmount' || inputMethod.value === 'outputNativeValue';
-    return isQuoteStale.value === 1 && isAdjustingOutputValue ? 1 : 0;
-  });
-
-  const isOutputStale = useDerivedValue(() => {
-    const isAdjustingInputValue =
-      inputMethod.value === 'inputAmount' || inputMethod.value === 'inputNativeValue' || inputMethod.value === 'slider';
-    return isQuoteStale.value === 1 && isAdjustingInputValue ? 1 : 0;
-  });
-
-  const pulsingOpacity = useDerivedValue(() => {
-    return isQuoteStale.value === 1
-      ? withRepeat(withSequence(withTiming(0.5, pulsingConfig), withTiming(1, pulsingConfig)), -1, true)
-      : withSpring(1, sliderConfig);
-  }, []);
-
-  const inputAmountTextStyle = useAnimatedStyle(() => {
-    const isInputZero =
-      (inputValues.value.inputAmount === 0 && inputMethod.value !== 'slider') ||
-      (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0);
-    const isOutputZero = Number(inputValues.value.outputAmount) === 0;
-
-    // eslint-disable-next-line no-nested-ternary
-    const zeroOrAssetColor = isInputZero ? zeroAmountColor : topColor === ETH_COLOR_DARK ? ETH_COLOR_DARK_ACCENT : topColor;
-    const opacity = isInputStale.value !== 1 || (isInputZero && isOutputZero) ? withSpring(1, sliderConfig) : pulsingOpacity.value;
-
-    return {
-      color: withTiming(interpolateColor(isInputStale.value, [0, 1], [zeroOrAssetColor, zeroAmountColor]), slowFadeConfig),
-      flexGrow: 0,
-      flexShrink: 1,
-      opacity,
-    };
-  }, [isDarkMode, topColor]);
-
-  const inputNativeValueStyle = useAnimatedStyle(() => {
-    const isInputZero =
-      Number(inputValues.value.inputAmount) === 0 || (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0);
-    const isOutputZero = Number(inputValues.value.outputAmount) === 0;
-
-    const zeroOrColor = isInputZero ? zeroAmountColor : labelTertiary;
-    const opacity = isInputStale.value !== 1 || (isInputZero && isOutputZero) ? withSpring(1, sliderConfig) : pulsingOpacity.value;
-
-    return {
-      color: withTiming(interpolateColor(isInputStale.value, [0, 1], [zeroOrColor, zeroAmountColor]), slowFadeConfig),
-      opacity,
-    };
-  }, [isDarkMode]);
-
-  const outputAmountTextStyle = useAnimatedStyle(() => {
-    const isInputZero =
-      Number(inputValues.value.inputAmount) === 0 || (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0);
-    const isOutputZero =
-      (inputValues.value.outputAmount === 0 && inputMethod.value !== 'slider') ||
-      (inputMethod.value === 'slider' && Number(inputValues.value.outputAmount) === 0);
-
-    // eslint-disable-next-line no-nested-ternary
-    const zeroOrAssetColor = isOutputZero ? zeroAmountColor : bottomColor === ETH_COLOR_DARK ? ETH_COLOR_DARK_ACCENT : bottomColor;
-    const opacity = isOutputStale.value !== 1 || (isInputZero && isOutputZero) ? withSpring(1, sliderConfig) : pulsingOpacity.value;
-
-    return {
-      color: withTiming(interpolateColor(isOutputStale.value, [0, 1], [zeroOrAssetColor, zeroAmountColor]), slowFadeConfig),
-      flexGrow: 0,
-      flexShrink: 1,
-      opacity,
-    };
-  }, [bottomColor, isDarkMode]);
-
-  const outputNativeValueStyle = useAnimatedStyle(() => {
-    const isInputZero =
-      Number(inputValues.value.inputAmount) === 0 || (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0);
-    const isOutputZero = Number(inputValues.value.outputAmount) === 0;
-
-    const zeroOrColor = isOutputZero ? zeroAmountColor : labelTertiary;
-    const opacity = isOutputStale.value !== 1 || (isInputZero && isOutputZero) ? withSpring(1, sliderConfig) : pulsingOpacity.value;
-
-    return {
-      color: withTiming(interpolateColor(isOutputStale.value, [0, 1], [zeroOrColor, zeroAmountColor]), slowFadeConfig),
-      opacity,
-    };
-  }, [isDarkMode]);
-
-  // TODO: Create a reusable InputCaret component
-  const inputCaretStyle = useAnimatedStyle(() => {
-    const shouldShow =
-      focusedInput.value === 'inputAmount' &&
-      inputProgress.value === 0 &&
-      outputProgress.value === 0 &&
-      (inputMethod.value !== 'slider' ||
-        (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0) ||
-        (sliderPressProgress.value === SLIDER_COLLAPSED_HEIGHT / SLIDER_HEIGHT && isQuoteStale.value === 0));
-
-    const opacity = shouldShow
-      ? withRepeat(
-          withSequence(
-            withTiming(1, { duration: 0 }),
-            withTiming(1, { duration: 400, easing: Easing.bezier(0.87, 0, 0.13, 1) }),
-            withTiming(0, caretConfig),
-            withTiming(1, caretConfig)
-          ),
-          -1,
-          true
-        )
-      : withTiming(0, caretConfig);
-
-    const isZero =
-      (inputMethod.value !== 'slider' && inputValues.value.inputAmount === 0) ||
-      (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0);
-
-    return {
-      display: shouldShow ? 'flex' : 'none',
-      opacity,
-      position: isZero ? 'absolute' : 'relative',
-    };
-  });
-
-  const outputCaretStyle = useAnimatedStyle(() => {
-    const shouldShow =
-      focusedInput.value === 'outputAmount' &&
-      inputProgress.value === 0 &&
-      outputProgress.value === 0 &&
-      (inputMethod.value !== 'slider' ||
-        (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0) ||
-        (sliderPressProgress.value === SLIDER_COLLAPSED_HEIGHT / SLIDER_HEIGHT && isQuoteStale.value === 0));
-
-    const opacity = shouldShow
-      ? withRepeat(
-          withSequence(
-            withTiming(1, { duration: 0 }),
-            withTiming(1, { duration: 400, easing: Easing.bezier(0.87, 0, 0.13, 1) }),
-            withTiming(0, caretConfig),
-            withTiming(1, caretConfig)
-          ),
-          -1,
-          true
-        )
-      : withTiming(0, caretConfig);
-
-    const isZero =
-      (inputMethod.value !== 'slider' && inputValues.value.outputAmount === 0) ||
-      (inputMethod.value === 'slider' && Number(inputValues.value.inputAmount) === 0);
-
-    return {
-      display: shouldShow ? 'flex' : 'none',
-      opacity,
-      position: isZero ? 'absolute' : 'relative',
-    };
-  });
-
-  return {
-    inputAmountTextStyle,
-    inputCaretStyle,
-    inputNativeValueStyle,
-    outputAmountTextStyle,
-    outputCaretStyle,
-    outputNativeValueStyle,
-  };
-}
-//
-// /---- END swap UI hooks ----/ //
